# backend/server/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Documentation VerhuurVanFietsen",
    "description": "A full documentation of the api VerhuurVanFietsen"
  },
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/bikeinuse/user/{id}": {
      "get": {
        "tags": ["BikeInUse"],
        "summary": "Get one bikeInUse item in system",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Find BikeInUse with userid",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BikeInUse"
            }
          },
          "404": {
            "description": "Item with 'id' not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/hiredhistory/user/{id}": {
      "get": {
        "tags": ["HiredHistory"],
        "summary": "Get one hiredhistory item in system",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Find hiredhistory with userid",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BikeInUse"
            }
          },
          "404": {
            "description": "Item with 'id' not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/all": {
      "get": {
        "tags": ["User"],
        "summary": "Get all user items in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "403": {
            "description": "You are not authorized to acces this route"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get one user item in system",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Find user with id",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Problem"
            }
          },
          "404": {
            "description": "Item with 'id' not found"
          },
          "403": {
            "description": "You are not authorized to acces this route"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete user item with given id",
        "tags": ["User"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Delete user item with id",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/id"
            }
          },
          "403": {
            "description": "You are not authorized to delete other then your own account"
          },
          "404": {
            "description": "User with 'id' not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/signup": {
      "post": {
        "tags": ["User"],
        "summary": "Create a new user item in system",
        "requestBody": {
          "description": "user Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/id"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
    "/signup/admin": {
      "post": {
        "tags": ["User"],
        "summary": "Create a new admin user item in system",
        "requestBody": {
          "description": "user Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/id"
            }
          },
          "403": {
            "description": "You are not authorized to acces this route"
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    }
  },
  "definitions": {
    "id": {
      "properties": {
        "id": {
          "type": "integer"
        }
      }
    },
...
    "BikeInUse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "bikeId": {
          "type": "integer"
        },
        "timeHired": {
          "type": "string",
          "format": "date"
        },
        "userId": {
          "type": "string"
        }
      }
    },
...
    "Problem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "bikeId": {
          "type": "integer"
        },
        "bikeStorageId": {
          "type": "string"
        },
        "date": {
          "type": "string",
          "format": "date"
        },
        "category": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageUrls": {
          "type": "string"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}